♻️[Refactor] 탭, 검색, 필터 구조를 FiterBar 공통 컴포넌트로 분리 #204
