+++ src/components/common-layout/SearchSection.tsx
