+++ src/components/common-layout/index.ts
@@ -1,0 +1,3 @@
+export { default as FilterSection } from './FilterSection';
+export { default as SearchSection } from './SearchSection';
+export { default as CategoryTabs } from './CategoryTabs';
