+++ src/features/bookclub/components/index.ts
@@ -1,5 +1,3 @@
 export { default as BookClubMainPage } from './BookClubMainPage';
 export { default as HeaderSection } from './HeaderSection';
 export { default as ClubListSection } from './ClubListSection';
-export { default as FilterSection } from './FilterSection';
-export { default as SearchSection } from './SearchSection';
